requirements.txt
python-telegram-bot
requests
ffmpeg-python
gTTS
python-dotenv
aiohttp
 Pillow
